Implementation of theia and policy recommendation CLI
In this commit, we add the implementation of theia, the command-line
tool for Theia based on cobra framework.
Policy recommendation CLI is also included. It consists of 3 commands
which let users could run a new policy recommendation job, check the
current status of a previous job, and retrieve the recommendation
result of a previous job respectively.
In order to access theia services externally, theia can
make use of k8s port forwarder functionality. Port forwarder
provides tooling to start/stop port forwarding channel for
eather Pod+Port or Service+Port. The code is loosely based on
corresponding code in kubectl.

Signed-off-by: Yongming Ding <dyongming@vmware.com>
Co-authored-by: Yanjun Zhou <zhouya@vmware.com>
Co-authored-by: Anna Khmelnitsky <akhmelnitsky@vmware.com>